+++ src/dev/ACS71240.cpp
@@ -4,12 +4,12 @@ namespace LVSS {
 
 ACS71240::ACS71240(io::ADC& adc0) : ADC(adc0)   {}
 
-int32_t ACS71240::readCurrent() {
+uint16_t ACS71240::readCurrent() {
     // Gets adcCounts from adc
-    int32_t adcCounts = ADC.readRaw();
+    int16_t adcCounts = ADC.readRaw();
 
     //(((adcCounts - average adc counts) * inputVoltage) / (sensitivity * 2^12)/100)
-    int32_t current = (((adcCounts - avgAdcCount) * voltIn) / (sensitivity << 12)/100);
+    int16_t current = (((adcCounts - avgAdcCount) * voltIn) / (sensitivity << 12)/100);
 
     return current;
 }
